VibeSync - V6/requirements.txt
kivy
deepface
pynput
windows-toasts
python-dotenv
requests
matplotlib
deepface
windows-toasts
pynput
python-dotenv
requests
pywin32
